- Changed variable for G_malloc() to size_t storage type.
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@17739 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ lib/display/raster_rgb.c
@@ -113,9 +113,9 @@ static int draw_cell_RGB(
     int r_size = G_raster_size(r_type);
     int g_size = G_raster_size(g_type);
     int b_size = G_raster_size(b_type);
-    int r_bytes = (D_x_end - D_x_beg) * r_size;
-    int g_bytes = (D_x_end - D_x_beg) * g_size;
-    int b_bytes = (D_x_end - D_x_beg) * b_size;
+    size_t r_bytes = (D_x_end - D_x_beg) * r_size;
+    size_t g_bytes = (D_x_end - D_x_beg) * g_size;
+    size_t b_bytes = (D_x_end - D_x_beg) * b_size;
 
 /* Allocate memory for raster */
     if(!r_raster)
